Modify chunk-related code to be more consistent and use size_t for array-related values

+++ src/value.h
@@ -84,8 +84,8 @@ typedef struct {
 #endif
 
 typedef struct {
-    uint32_t count;
-    uint32_t capacity;
+    size_t count;
+    size_t capacity;
     Value* values;
 } ValueArray;
